Fixed insidious OpenCV mat access bug.

+++ src/Homography33.cc
@@ -175,7 +175,7 @@ void Homography33::compute() {
   cv::Mat homography = cv::findHomography(sPts, dPts);
   for (int c=0; c<3; c++) {
     for (int r=0; r<3; r++) {
-      H(r,c) = homography.at<float>(r,c);
+      H(r,c) = homography.at<double>(r,c);
     }
   }
